fix: keep back-to-top button above footer
