+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import { type ReactNode } from "react"
 import { ThemeProvider } from "@wrksz/themes/next"
 
 import { Analytics } from "@/components/analytics"
-import { BackToTop } from "@/components/site/back-top-top"
+import { BackToTop } from "@/components/site/back-to-top"
 import { SiteFooter } from "@/components/site/site-footer"
 import { SiteHeader } from "@/components/site/site-header"
 import { TooltipProvider } from "@/components/ui/shadcn/tooltip"
